requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-api-key==3.0.0
geographiclib==2.0
geopy==2.4.1
idna==3.6
packaging==23.2
python-decouple==3.8
pytz==2023.4
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
tzdata==2023.4
urllib3==2.1.0
